Customize FFMPEG to support DJI SDK video decoding

+++ libavcodec/huffyuvencdsp.h
@@ -25,13 +25,13 @@ typedef struct HuffYUVEncDSPContext {
     void (*diff_bytes)(uint8_t *dst /* align 16 */,
                        const uint8_t *src1 /* align 16 */,
                        const uint8_t *src2 /* align 1 */,
-                       intptr_t w);
+                       int w);
     /**
      * Subtract HuffYUV's variant of median prediction.
      * Note, this might read from src1[-1], src2[-1].
      */
     void (*sub_hfyu_median_pred)(uint8_t *dst, const uint8_t *src1,
-                                 const uint8_t *src2, intptr_t w,
+                                 const uint8_t *src2, int w,
                                  int *left, int *left_top);
 } HuffYUVEncDSPContext;
